requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece
jieba
pandas
numpy
scikit-learn
wandb
tqdm
pyyaml
shap==0.50.0
tokenizers==0.22.1
regex==2025.9.18
scipy==1.16.2





